requirement.txt
adbutils==0.14.1
allure-pytest==2.9.45
allure-python-commons==2.9.45
altgraph==0.17.2
apkutils2==1.0.0
APScheduler==3.10.1
astroid==2.11.2
atomicwrites==1.4.0
attrs==21.4.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.12
cigam==0.0.3
colorama==0.4.4
decorator==4.4.2
Deprecated==1.2.13
deprecation==2.1.0
dill==0.3.4
facebook-wda==1.4.6
filelock==3.6.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
imageio==2.8.0
imageio-ffmpeg==0.4.5
importlib-metadata==4.11.3
iniconfig==1.1.1
isort==5.10.1
JPype1==1.0.2
lazy-object-proxy==1.7.1
logzero==1.7.0
lxml==4.8.0
mccabe==0.7.0
moviepy==1.0.3
numpy==1.21.5
opencv-contrib-python==3.4.2.17
packaging==20.9
pefile==2021.9.3
Pillow==9.1.0
platformdirs==2.5.1
pluggy==1.0.0
proglog==0.1.10
progress==1.6
psutil==5.9.5
py==1.11.0
pyelftools==0.28
pygame==2.5.2
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.3
pyparsing==3.0.7
pytest==7.1.1
pytest-assume==2.4.3
pytest-dependency==0.5.1
pytest-ordering==0.6
pytest-random-order==1.1.0
pytest-repeat==0.9.1
pytest-rerunfailures==10.2
pytest_check==1.0.5
pytz==2023.3
pywin32==303
pywin32-ctypes==0.2.0
PyYAML==6.0
qrcode==7.3.1
requests==2.27.1
retry==0.9.2
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
tomli==2.0.1
tornado==6.1
tqdm==4.66.1
typed-ast==1.5.2
typing_extensions==4.1.1
tzdata==2023.3
tzlocal==5.0.1
uiautomator2==2.16.14
urllib3==1.26.9
websocket-client==0.57.0
whichcraft==0.6.1
wrapt==1.14.0
xmltodict==0.12.0
zipp==3.8.0
